+++ src/oersted/__init__.py
@@ -5,9 +5,10 @@ from .solver import Solver, DirectSolver, OctreeSolver
 
 from .biotsavart import b_field, h_field, h_mag
 
-from .mesh import Mesh, CentroidMesh, mesh_step, plot_mesh, surface_forces
+from .mesh import Mesh, CentroidMesh, mesh_step, plot_mesh
 from . import testing
 from . import magnetization
+from .results import kelvin_force_density, kelvin_forces, maxwell_forces
 
 
 __all__ = [
@@ -32,5 +33,7 @@ __all__ = [
     "DirectSolver",
     "OctreeSolver",
     "magnetization",
-    "surface_forces",
+    "maxwell_forces",
+    "kelvin_forces",
+    "kelvin_force_density",
 ]