+++ tests/test_surface_forces.py
@@ -1,54 +1,40 @@
 import numpy as np
 import oersted
-from oersted import Mesh, MU0, OctreeSolver
-import time
+from oersted import Mesh
 
 # Test parameters
-infile: str = "tests/data/sphere.stp"
-mesh_size: float = 15  # mm
-b_ext_mag: float = 1.0  # T
+mesh_size: float = 15.0  # mm
+b_ext_mag: float = 5.0  # T
 mu_r: float = 1.5
 solver = oersted.DirectSolver()
 mat = oersted.materials.LinearMaterial(mu_r)
 
-# Mesh the sphere
-mesh: Mesh = oersted.mesh.mesh_step(infile, mesh_size, mesh_size)
-print(f"Number of elements: {mesh.num_elems}")
+def calculate_magnetization_forces_on_sphere(b_ext_mag: float, b_ext_grad: float):
+    """Test maxwell forces for a magnetized component under an external field
 
-mesh = oersted.Mesh(mesh.nodes, mesh.connectivity)
-print(np.average(mesh.centroids, axis=0))
+    Assume that the gradient in the field is in the z-direction"""
 
+    # Mesh the sphere
+    mesh: Mesh = oersted.mesh.mesh_step("tests/data/sphere.stp", mesh_size, mesh_size)
+    mesh.nodes[:, 2] -= 1.0
 
-def test_magnetization_forces():
-    """Test maxwell forces for a magnetized component under an external field"""
-
-    # Calculate uniform background field (force should be near-zero)
+    # Calculate background field
     h_external = np.zeros((mesh.num_elems, 3))
-    h_ext_mag: float = b_ext_mag / oersted.MU0
-    h_external[:, 2] = h_ext_mag
+    h_external[:, 2] = (1.0 / oersted.MU0) * (b_ext_mag + b_ext_grad * mesh.centroids[:, 2])
 
     # Compute demag parameters: magnetization and internal H field
-    start = time.perf_counter()
-    M, Htotal = oersted.magnetization.demag_tet4(mesh, mat, h_external, nthreads_requested=solver.n_threads, theta=0.5, leaf_threshold=16)
-    elapsed = time.perf_counter() - start
-    print(f"Calculation time elapsed: {elapsed:.3f} sec")
+    M, Htotal = oersted.magnetization.demag_tet4(mesh, mat, h_external, solver)
 
-    # Compute external field at mesh face centroids
-    b_ext = np.zeros(mesh.surface.centroids.shape)
-    b_ext[:, 2] = b_ext_mag
-    offset = 1e-4  # small distance outward
-    eval_pts = mesh.surface.centroids + offset * mesh.surface.normals
-    # h_demag = oersted.magnetization.h_demag_tet4(mesh,M, eval_pts)
-    h_demag = oersted.biotsavart.h_mag(mesh, M, eval_pts, solver=OctreeSolver(0.5, 16, 0))
-    b_ext = oersted.MU0 * (b_ext / MU0 + h_demag)
-    forces = oersted.mesh.surface_forces(mesh.surface, b_ext, solver)
+    # Compute external field at mesh nodes
+    h_ext_nodes = np.zeros((mesh.num_nodes, 3))
+    h_ext_nodes[:, 2] = (1.0 / oersted.MU0) * (b_ext_mag + b_ext_grad * mesh.nodes[:, 2])
+    forces = oersted.kelvin_forces(mesh, M, h_ext_nodes)
 
-    total_force = np.sum(forces, axis=0)
-    print(np.sum(forces, axis=0))
+    # Compute analytical result: f = (m*grad)H
+    m = np.average(M, axis=0)[2] * np.sum(mesh.volumes)
+    analytical_force = m * b_ext_grad / oersted.MU0
 
-    print(f"Uniform field force: {np.sum(forces, axis=0)}")
-
-    assert np.linalg.norm(total_force) < 5.0  # small value
+    return np.sum(forces, axis=0), analytical_force
 
 
 def test_lorentz_forces():
@@ -57,7 +43,7 @@ def test_lorentz_forces():
     # Make the helmholz coil problem
     radius = 0.2
     total_current = 1e4
-    mesh_size: float = 10.0
+    mesh_size: float = 15.0
     mesh1 = oersted.mesh_step("tests/data/ring.stp", mesh_size, mesh_size)
     mesh1._nodes[:, 2] += 0.01
 
@@ -84,10 +70,29 @@ def test_lorentz_forces():
     bext = oersted.b_field(mesh1, jdensity, mesh2.surface.centroids, solver=solver)
     bext += oersted.b_field(mesh2, jdensity, mesh2.surface.centroids, solver=solver)
 
-    forces = oersted.mesh.surface_forces(mesh2.surface, bext, solver)
+    forces = oersted.maxwell_forces(mesh2.surface, bext)
     total_force = np.sum(forces, axis=0)
-    print(total_force)
-    assert np.abs((fz_expected - total_force[2]) / fz_expected) < 1e-2
+    error = np.abs((fz_expected - total_force[2]) / fz_expected)
+    print(f"total force, lower: {total_force}")
+
+    print(f"error: {100 * error:.2f} %")
+    assert error < 1e-2
+
+    # Check that the other components are small, like less than 1.0 N
+    assert np.abs(total_force[0]) < 1.0
+    assert np.abs(total_force[1]) < 1.0
+
+    # Compute the field at the lower coil's surface elements using both coils
+    bext = oersted.b_field(mesh1, jdensity, mesh1.surface.centroids, solver=solver)
+    bext += oersted.b_field(mesh2, jdensity, mesh1.surface.centroids, solver=solver)
+
+    forces = oersted.maxwell_forces(mesh1.surface, bext)
+    total_force = np.sum(forces, axis=0)
+    error = np.abs((-fz_expected - total_force[2]) / fz_expected)
+    print(f"total force, upper: {total_force}")
+
+    print(f"error: {100 * error:.2f} %")
+    assert error < 1e-2
 
     # Check that the other components are small, like less than 1.0 N
     assert np.abs(total_force[0]) < 1.0
@@ -95,8 +100,27 @@ def test_lorentz_forces():
 
 
 def main():
-    test_magnetization_forces()
     test_lorentz_forces()
+
+    # Magnetized sphere in uniform background field
+    b_ext_mag = 5.0
+    b_ext_grad = 0.0
+    total_force, analytical_force = calculate_magnetization_forces_on_sphere(b_ext_mag, b_ext_grad)
+    print(f"total:      {total_force}")
+    print(f"analytical: {analytical_force}")
+    assert np.abs(total_force[0]) < 1e-8
+    assert np.abs(total_force[1]) < 1e-8
+    assert np.abs(total_force[2] - analytical_force) < 1e-8  # expected value is zero
+
+    # Magnetized sphere in linear background field gradient
+    b_ext_mag = 5.0
+    b_ext_grad = 1.0
+    total_force, analytical_force = calculate_magnetization_forces_on_sphere(b_ext_mag, b_ext_grad)
+    print(f"total:      {total_force}")
+    print(f"analytical: {analytical_force}")
+    assert np.abs(total_force[0]) < 1e-8
+    assert np.abs(total_force[1]) < 1e-8
+    assert np.abs((total_force[2] - analytical_force) / analytical_force) < 1e-4
 
 
 if __name__ == "__main__":